[util] Factor out common monitor-related functions
