+++ src/d3d9/d3d9_include.h
@@ -36,8 +36,9 @@
 #include "../util/util_flags.h"
 #include "../util/util_likely.h"
 #include "../util/util_math.h"
-#include "../util/util_string.h"
+#include "../util/util_monitor.h"
 #include "../util/util_misc.h"
+#include "../util/util_string.h"
 
 // Missed definitions in Wine/MinGW.
