+++ src/dxgi/dxgi_include.h
@@ -25,6 +25,7 @@
 #include "../util/util_flags.h"
 #include "../util/util_likely.h"
 #include "../util/util_math.h"
+#include "../util/util_monitor.h"
 #include "../util/util_string.h"
 
 #include <dxgi1_5.h>
